metal2(post-port): address review on data_movement self-loop conversions
Two review fixes:
- copy: the reader-private scratchpad only takes get_base_address(), so use
  Scratchpad<uint8_t>, not <uint32_t>. DefaultRowMajor page sizes need not be
  divisible by 4 (UINT8/FP8, odd row-major widths), which would break the
  Scratchpad size%sizeof(T) invariant; a byte view is valid for any page size
  (recipe T-table: address-only -> uint8_t).
- scatter: repair the reader comment falsified by this PR -- FP32_TEMP is a
  Scratchpad now, no longer a self-loop DFB with PRODUCER/CONSUMER bindings.

Behaviour-preserving; test_copy.py (542 passed) and test_scatter.py -k reduction
(19 passed) green before and after.

Relates to #55526.

+++ ttnn/cpp/ttnn/operations/data_movement/copy/device/kernels/redistribute_pages_row_major_reader.cpp
@@ -35,7 +35,9 @@ void kernel_main() {
     Noc noc;
     // scratch::in0 is a reader-private L1 scratchpad: the reader stages input pages in it.
     // dfb::in1 is the reader->writer output-page FIFO.
-    Scratchpad<std::uint32_t> in0(scratch::in0);
+    // uint8_t view: the kernel only takes the base address; a byte element type keeps the
+    // Scratchpad size%sizeof(T) invariant valid for any page size (UINT8/FP8, odd RM widths).
+    Scratchpad<std::uint8_t> in0(scratch::in0);
     DataflowBuffer dfb_in1(dfb::in1);
 
     const auto accessor_src = TensorAccessor(tensor::src);
